refactor: move all Header components into a header folder
