+++ src/components/RootLayout.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image"
 import Link from "next/link"
-import Header from "@components/Header"
+import Header from "@components/Header/Header"
 import Footer from "@components/Footer"
 import Main from "@components/Main"
